+++ lowdown.h
@@ -179,7 +179,7 @@ struct	lowdown_node {
 		} rndr_list; 
 		struct rndr_listitem {
 			enum hlist_fl flags; /* all possible flags */
-			uint32_t num; /* index in ordered */
+			size_t num; /* index in ordered */
 		} rndr_listitem; 
 		struct rndr_header {
 			size_t level; /* hN level */
